refactor: move entity internals closer to owners
