+++ src/pages/dead-code/widgets/DeadCodeExplorer/ui/DeadCodeCategory.tsx
@@ -9,7 +9,6 @@ import type React from 'react';
 import { useMemo } from 'react';
 import { activeActivityPageIdAtom } from '@/app/model/activityPageAtoms';
 import { filesAtom, viewModeAtom } from '@/entities/AppView/model/atoms';
-import { FileIcon } from '@/entities/SourceFileNode/ui/FileIcon.tsx';
 import { buildDeadCodeTree } from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/lib/buildDeadCodeTree.ts';
 import { renderCategoryIcon } from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/lib/categoryUtils.tsx';
 import type { DeadCodeItem } from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/lib/deadCodeAnalyzer.ts';
@@ -23,6 +22,7 @@ import { CategoryCheckbox } from '@/features/Code/CodeAnalyzer/DeadCodeSelection
 import { targetLineAtom } from '@/features/File/Navigation/model/atoms.ts';
 import { useOpenFile } from '@/features/File/OpenFiles/lib/useOpenFile.ts';
 import { Checkbox } from '@/shared/ui/Checkbox';
+import { FileIcon } from '@/shared/ui/FileIcon';
 import { FileTreeItem } from '@/shared/ui/FileTreeItem';
 import { TreeView } from '@/shared/ui/TreeView/TreeView.tsx';
 import { DeadCodeFolderItem } from './DeadCodeFolderItem.tsx';
