+++ src/pages/dead-code/widgets/DeadCodeExplorer/ui/DeadCodeFileItem.tsx
@@ -6,12 +6,12 @@ import { useSetAtom } from 'jotai';
 import React from 'react';
 import { activeActivityPageIdAtom } from '@/app/model/activityPageAtoms';
 import { viewModeAtom } from '@/entities/AppView/model/atoms';
-import { FileIcon } from '@/entities/SourceFileNode/ui/FileIcon.tsx';
 import type { DeadCodeItem } from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/lib/deadCodeAnalyzer.ts';
 import { useDeadCodeSelection } from '@/features/Code/CodeAnalyzer/DeadCodeSelection/lib/useDeadCodeSelection.ts';
 import { targetLineAtom } from '@/features/File/Navigation/model/atoms.ts';
 import { useOpenFile } from '@/features/File/OpenFiles/lib/useOpenFile.ts';
 import { Checkbox } from '@/shared/ui/Checkbox';
+import { FileIcon } from '@/shared/ui/FileIcon';
 import { FileTreeItem } from '@/shared/ui/FileTreeItem';
 
 interface DeadCodeFileItemRowProps {
