+++ src/shared/ui/FileIcon.tsx
