+++ src/widgets/FileExplorer/FileExplorer.tsx
@@ -10,8 +10,8 @@ import { useEffect, useMemo, useState } from 'react';
 import { filesAtom, focusedFolderAtom } from '@/entities/AppView/model/atoms';
 import { useOpenFile } from '@/features/File/OpenFiles/lib/useOpenFile';
 import { activeTabAtom, openedTabsAtom } from '@/features/File/OpenFiles/model/atoms';
+import { FileIcon } from '@/shared/ui/FileIcon';
 import { FileTreeItem } from '@/shared/ui/FileTreeItem';
-import { FileIcon } from '../../entities/SourceFileNode/ui/FileIcon.tsx';
 import { useTreeKeyboardNavigation } from '../../shared/hooks/useTreeKeyboardNavigation';
 import { TreeView } from '../../shared/ui/TreeView/TreeView';
 import { buildFileTree } from './lib/buildFileTree';
