+++ src/widgets/MainContents/IDEScrollView/ui/FileSection.tsx
@@ -7,8 +7,8 @@ import { useAtomValue, useSetAtom } from 'jotai';
 import { forwardRef, useEffect, useRef, useState, useTransition } from 'react';
 import { deadCodeResultsAtom } from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/model/atoms.ts';
 import { activeTabAtom } from '@/features/File/OpenFiles/model/atoms.ts';
+import { FileIcon } from '@/shared/ui/FileIcon';
 import type { SourceFileNode } from '../../../../entities/SourceFileNode/model/types.ts';
-import { FileIcon } from '../../../../entities/SourceFileNode/ui/FileIcon.tsx';
 import { getFileName } from '../../../../shared/pathUtils.ts';
 import { getWorkerPool } from '../../../../shared/workerPool.ts';
 import CodeViewer from '../../../CodeViewer/CodeViewer.tsx';
